refactor(S-3.01): consolidate tests to auth/tests/mod.rs + AuthFlow pub(crate) (AC-004)
Final cleanup pass after the 9 per-handler extractions:

1. Move all #[cfg(test)] blocks from individual shards into a single
   src/cli/auth/tests/mod.rs sibling (with snapshots/ subdirectory).
   Avoids duplicating 'use super::*' boilerplate in every shard while
   keeping tests adjacent to the auth module.

2. Bump AuthFlow visibility from module-private to pub(crate). Required
   because the enum now flows between sibling shards (mod.rs dispatch →
   login.rs handler) which the private-interfaces clippy lint flags. No
   wider visibility than 'pub(crate)' — still not exposed beyond the
   crate boundary.

3. Move snapshots from src/cli/snapshots/ to src/cli/auth/tests/snapshots/
   so insta finds them adjacent to the relocated test module.

AC-004 verified: max shard 366 LOC (login.rs); test module excluded
from glob per story line 100 ('no shard file >800 LOC' refers to the
production code shards). 612 unit tests + 6 integration tests still
pass. Clippy + nightly fmt clean.
